+++ packages/php/livewire/tests/TestCase.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Tests;
+
+use MrPunyapal\ClientValidation\ClientValidationServiceProvider;
+use Orchestra\Testbench\TestCase as Orchestra;
+
+class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app)
+    {
+        return [
+            ClientValidationServiceProvider::class,
+        ];
+    }
+
+    public function getEnvironmentSetUp($app)
+    {
+        config()->set('database.default', 'testing');
+    }
+}